requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
celery>=5.4,<5.6
redis>=5.0,<6.0
dnspython>=2.6,<3.0
httpx>=0.27,<1.0
tenacity>=9.0,<10.0
python-json-logger>=2.0,<3.0
dj-database-url>=2.2,<3.0
psycopg[binary]>=3.2,<4.0
black>=24.0,<25.0
ruff>=0.6,<1.0
